moved function to util
